fix parse_rates and remake signup

+++ src/account/signals.py
@@ -8,4 +8,5 @@ from currency_exchange.settings import MEDIA_ROOT
 
 @receiver(pre_save, sender=User)
 def pre_save_User(sender, instance, **kwargs):
-    shutil.rmtree(os.path.join(MEDIA_ROOT, 'avatar', str(instance.id)))
+    if instance.avatar != None:
+        shutil.rmtree(os.path.join(MEDIA_ROOT, 'avatar', str(instance.id)))
